Demonstrate issue with KFAC with bf16 precision
Currently, KFAC does not work well when run with --precision bf16, but the
existing tests are not well-suited to understanding and fixing the issue
properly. This commit therefox fixes two issues:

- we change the ground truth computation of eigenvectors and eigenvalues
to use float32 vectors and operations even if we're running the model itself in
bf16, this way it stays close to the actual ground truth we care about. This
avoids mistakenly thinking the real computation is good enough when it isn't.

- We replace the fragile test_eigenvectors and test_eigenvalue_correction by a
new test_ekfac_recomposition. This is more robust because the eigendecomposition
itself can be very different due to small numerical differences even if the
recomposed matrices are very close. This avoids mistakenly thinking the real
computation is not good enough when it is.

After these two changes, we observe that bf16 support is indeed broken:

$ pytest tests/ekfac_tests/test_ekfac_recomposition.py -sv --precision bf16 --model_name EleutherAI/Pythia-14m

FAILED tests/ekfac_tests/test_ekfac_recomposition.py::test_ekfac_recomposition - AssertionError: EKFAC GNH recomposition: max per-layer rel_error=2.66e-01, expected < 0.001
    layers.0.attention.dense: rel_error=7.17e-02
    layers.0.attention.query_key_value: rel_error=5.46e-02
    layers.0.mlp.dense_4h_to_h: rel_error=5.56e-02
    layers.0.mlp.dense_h_to_4h: rel_error=1.05e-01
    layers.1.attention.dense: rel_error=6.45e-02
    layers.1.attention.query_key_value: rel_error=8.21e-02
    layers.1.mlp.dense_4h_to_h: rel_error=5.80e-02
    layers.1.mlp.dense_h_to_4h: rel_error=6.93e-02
    layers.2.attention.dense: rel_error=3.20e-02
    layers.2.attention.query_key_value: rel_error=0.00e+00
    layers.2.mlp.dense_4h_to_h: rel_error=8.45e-02
    layers.2.mlp.dense_h_to_4h: rel_error=6.08e-02
    layers.3.attention.dense: rel_error=5.45e-02
    layers.3.attention.query_key_value: rel_error=2.66e-01
    layers.3.mlp.dense_4h_to_h: rel_error=6.06e-02
    layers.3.mlp.dense_h_to_4h: rel_error=3.36e-02
    layers.4.attention.dense: rel_error=2.04e-02
    layers.4.attention.query_key_value: rel_error=9.69e-02
    layers.4.mlp.dense_4h_to_h: rel_error=3.93e-02
    layers.4.mlp.dense_h_to_4h: rel_error=4.41e-02
    layers.5.attention.dense: rel_error=6.16e-02
    layers.5.attention.query_key_value: rel_error=1.79e-01
    layers.5.mlp.dense_4h_to_h: rel_error=3.09e-02
    layers.5.mlp.dense_h_to_4h: rel_error=5.30e-02
assert 0.266264408826828 < 0.001

+++ tests/ekfac_tests/compute_ekfac_ground_truth.py
@@ -584,8 +584,12 @@ def compute_eigenvectors_step(
             f"{name}: eigenvectors_a.sum()={eigenvectors_a.sum()}, "
             f"eigenvectors_g.sum()={eigenvectors_g.sum()}"
         )
-        eigenvectors_activations[name] = eigenvectors_a.to(dtype=dtype).contiguous()
-        eigenvectors_gradients[name] = eigenvectors_g.to(dtype=dtype).contiguous()
+        eigenvectors_activations[name] = eigenvectors_a.to(
+            dtype=torch.float32
+        ).contiguous()
+        eigenvectors_gradients[name] = eigenvectors_g.to(
+            dtype=torch.float32
+        ).contiguous()
 
     save_file(
         eigenvectors_activations,
